requirements.txt
# Data collection & processing
pandas>=2.0
requests>=2.31
pyarrow>=14.0

# NLP (Phase 2)
spacy>=3.7
vaderSentiment>=3.3
scikit-learn>=1.3

# Geocoding (Phase 2)
geopy>=2.4

# Scoring (Phase 3)
scipy>=1.11
numpy>=1.24
imbalanced-learn>=0.11
seaborn>=0.13
matplotlib>=3.8

# API (Phase 4)
fastapi>=0.104
uvicorn>=0.24
pymongo>=4.6
motor>=3.3

# Frontend support
python-dotenv>=1.0

# Reddit scraping
firecrawl-py>=1.0
